% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIPinnipedAnalysis-package.R
\docType{data}
\name{anchovy}
\alias{anchovy}
\title{Anchovy Biomass Estimates}
\format{A dataframe with fields 
\preformatted{
Year           - year 
AnchovyBiomass - estimate of stock biomass 1000 metric tons for the year
}}
\description{
A record for 1972-2011 from
MacCall, A. D., W. J. Sydeman, P. C. Davison, and J. A. Thayer. 2016. Recent collapse of northern anchovy biomass off California. Fish. Res. 175:87-94.
provided as file from Bill Sydeman
}
\keyword{datasets}
