% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_anomalies.R
\name{average_anomalies}
\alias{average_anomalies}
\title{Create average anomalies for a range of months}
\usage{
average_anomalies(v, start, n)
}
\arguments{
\item{v}{matrix of environmental values with columns for months and rows for years}

\item{start}{first location to use when v is treated as a vector of the matrix transposed}

\item{n}{number of months to include in each average}
}
\value{
vector of averages with year names
}
\description{
Create average anomalies for a range of months
}
\author{
Jeff Laake
}
