% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fo_data.r
\name{bone_cf}
\alias{bone_cf}
\title{Computes correction factor for pre-1992 data}
\usage{
bone_cf(fo)
}
\arguments{
\item{fo}{dataframe of prey species identified in each scat and whether it was identified by otolith/beak or other}
}
\value{
list with correction factor cf and its std error se.cf
}
\description{
For a given set of species codes it computes the correction factor for
oto/bk only data collected before 1992. It computes FO with all structures divided by
FO from oto/bk only and that ratio is then used to correct years prior to 1992 in
the fo function.
}
\author{
Jeff Laake
}

