% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIPinnipedAnalysis-package.R
\docType{data}
\name{calcofi}
\alias{calcofi}
\title{Calcofi Environmental Data Measurements}
\format{A dataframe with fields 
\preformatted{
Month - either July or October
Year - year of data measurement
Station - station value
dynamic_height_0_500m - dynamic heigth measurement - units unknown; function of temp and salinity
stratification 
pycnocline_depth  
R_POTEMP_25m - temperature at 25 meters
R_POTEMP_75m - temperature at 75 meters
R_SIGMA_25m - density gradient at 25m
R_SIGMA_75m - density gradient at 75m
R_O2_25m - oxygen at 25m
R_O2_75m - oxygen at 75m
R_NO3_25m - NO3 at 25m
R_NO3_75m - NO3 at 75m
R_SIGMA_75m - density gradient at 75m}}
\description{
A sequence of oceanographic measurements at seven calcofi stations in July from 1984-2013
Data were provided by Isaac Schroeder.
}
\keyword{datasets}

