% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upwelling.R
\name{compute.cumuwi}
\alias{compute.cumtodate}
\alias{compute.cumuwi}
\alias{plot_cumuwi}
\alias{plot_cumuwi.day}
\title{Upwelling functions}
\usage{
compute.cumuwi(uwi)
       plot_cumuwi(uwi,main="",colors=NULL)
       compute.cumtodate(cumuwi,day=NULL)
       plot_cumuwi.day(uwi,day=NULL,day2=NULL,main="")
}
\arguments{
\item{uwi}{dataframe of upwelling (see uwi33,uwi36,uwi39)}

\item{cumuwi}{cummulative uwi vector}

\item{main}{title for plot}

\item{colors}{vector of colors for cummulative upwelling lines}

\item{day}{julian day value at which cummulative upwelling is extracted by year}

\item{day2}{if not null then cummulative upwelling plotted from day to day2}
}
\value{
cum upwelling list from compute.cumuwi, and vector of cummulative 
upwelling values by year from compute.cumtodate
}
\description{
Upwelling functions
}
\examples{
{
 data(uwi33)
 data(uwi36)
 data(uwi39)
par(mfrow=c(3,1))
colors=rep("grey",2016-1967+1)
colors[c(2012,2014:2015)-1967+1]="green"
plot_cumuwi(uwi39,main="39N",colors=colors)	
plot_cumuwi(uwi36,main="36N",colors=colors)	
plot_cumuwi(uwi33,main="33N",colors=colors)	

par(mfrow=c(3,1))
plot_cumuwi.day(uwi39,main="39N End of Year")	
plot_cumuwi.day(uwi36,main="36N End of Year")	
plot_cumuwi.day(uwi33,main="33N End of Year")	

par(mfrow=c(3,1))
plot_cumuwi.day(uwi39,day=151,main="39N to 1 June")	
plot_cumuwi.day(uwi36,day=151,main="36N to 1 June")	
plot_cumuwi.day(uwi33,day=151,main="33N to 1 June")	

par(mfrow=c(3,1))
plot_cumuwi.day(uwi39,day=90,day2=151,main="39N 1 April to 1 June")	
plot_cumuwi.day(uwi36,day=90,day2=151,main="36N 1 April to 1 June")	
plot_cumuwi.day(uwi33,day=90,day2=151,main="33N 1 April to 1 June")	

}
}
\author{
Jeff Laake
}

