% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed.model.r
\name{compute_AICc}
\alias{compute_AICc}
\title{Compute AICc (AIC with small sample size correction) for a series of models and model.table with
number of parameters and model weight based on AICc rather than AIC.}
\usage{
compute_AICc(results, nobs, nref, fixed.f)
}
\arguments{
\item{results}{list of model results with element model.table}

\item{nobs}{number of observations}

\item{nref}{number of random effect parameters}

\item{fixed.f}{vector of fixed effect formulas}
}
\value{
list with following elements: results - list with each fitted model object; model.table- model selection table;
best - model number that was best fit in results list; best.f - formula for fixed effect from best model; best.r - formula for random effects for best model
}
\description{
Compute AICc (AIC with small sample size correction) for a series of models and model.table with
number of parameters and model weight based on AICc rather than AIC.
}
\author{
Jeff Laake
}
