% Generated by roxygen2 (4.0.2): do not edit by hand
\name{correct_dead}
\alias{correct_dead}
\title{Correction factors for dead pup counts}
\usage{
correct_dead(island, year, cfyear, cfdata, ndays = NULL)
}
\arguments{
\item{island}{("SMI" or "SNI")}

\item{year}{four digit numeric year}

\item{cfyear}{the year of tagging data used to create correction factor}

\item{ndays}{ndays is number of days past 1 July; if NULL provides a corrected value for each survey date (occasion); if non-null provides correction at that date}

\item{cfyears}{vector of years of tagging data used to create average correction factor}
}
\value{
a list containing a dataframe with corrections by strata (eg group - area or substrate/position) and a dataframe with a total across strata.
}
\description{
Correction factors for a set of dead pup counts are created for each survey date or a single date from
estimates created from POPAN model (cfdata) on tagging data.
}
\details{
For a given island and year, it calls getdead_ch  and process_ch to create the capture history for the dead pup data and then
uses process.data from RMark to process the capture history data for the POPAN model. If cfyear > 1997 it uses Position and Substrate as group factors and if
cfyear < 1997 it uses Area code which is split into 4 areas: PTS(NEP,NWP,PBP),SCV (SCV,PBS,EAC and WAC), WCV and NWC. The argument cfdata
contains the estimates from the POPAN model to create the correction factor. It is created by the function popan.cf which runs a set of POPAN models
with the tagging data from year yyyy (1994,1995,1998,2002 for SMI and 2006 for SNI) and creates smiyyyy.popan.results or sniyyyypopan.results with the model results and model averaged values for cfdata. If ndays is null it creates a correction factor for each
survey occasion. If ndays, number of days from 1 July, is not null then it computes with interpolation the number observed dead at that date and
its correction factor.

The function average_cf computes an average correction factor for a set of years with tagging data. This is only useful at present with SMI becuse SNI has only
one year of tagging data but it can be used with SNI as well.
}
\author{
Jeff Laake
}
\seealso{
popan.cf getdead_ch
}

