% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fo_data.r
\name{create_cufo}
\alias{create_cufo}
\title{Creates Frequency of Occurrence (FO) DataFrame for Primary Prey of Callorhinus}
\usage{
create_cufo(fo)
}
\arguments{
\item{fo}{dataframe of fo data included in package}
}
\value{
dataframe of FO for key prey or prey groupings
}
\description{
Creates Frequency of Occurrence (FO) DataFrame for Primary Prey of Callorhinus
}
\author{
Jeff Laake
}

