% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fo_data.r
\name{create_fo}
\alias{create_fo}
\title{Creates Frequency of Occurrence (FO) DataFrame for Primary Prey}
\usage{
create_fo(fo)
}
\arguments{
\item{fo}{dataframe of fo data included in package}
}
\value{
dataframe of FO for key prey or prey groupings and derived values
including first and second principal components of 5 primary prey
}
\description{
Creates Frequency of Occurrence (FO) DataFrame for Primary Prey
}
\author{
Jeff Laake
}

