% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIPinnipedAnalysis-package.R
\name{Cu}
\alias{Cu}
\title{Cu production and early mortailty}
\description{
Cu production and early mortailty
}
\examples{
\donttest{
 check.dead.pup.areas()
 source(file.path(system.file(package="CIPinnipedAnalysis"),"CuPupProduction.r"))
 source(file.path(system.file(package="CIPinnipedAnalysis"),"CuEarlyPupMortality.r"))
}
}
