% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popan.cf.r
\name{cu_popan.cf}
\alias{cu_popan.cf}
\title{Construct correction factor data from Cu dead pup tagging studies}
\usage{
cu_popan.cf(year, silent = TRUE)
}
\arguments{
\item{year}{four digit numeric year}

\item{silent}{if TRUE shows each model as it is run and any problems that occur; if FALSE this is hidden each survey date (occasion); if non-null provides correction at that date}
}
\value{
a list containing a marklist of the final model results, a list with the correction factor data (cfdata) and correction factors by occasion (cfbyocc).
}
\description{
The dead pup tagging data is analyzed with a set of POPAN 
models to create correction factors for the observed dead pup counts to estimate the total number that died.
}
\details{
The function cu_getdead_ch is used to extract the data from the CIPinnipedCensusMaster database and 
create capture histories for tagged and untagged (stacked) pups. This is then analyzed with the POPAN model in MARK using
the RMark package.  A set of models is examined that depends on the year because the number of sampling occasions
and the data collected has varied.  In POPAN models, p on the first and last occasion are not separately estimable in a time model
so these have been restricted such that p for occasion 1 is the same as occasion 2 and p for the last occasion is the same as p for the 
next to last occasion (see definition of time2).

Each set of models is run and then they are re-run using the initial values from the best model.  This helps with
model convergence which can fail for POPAN models.  To see which models are failing set silent=FALSE. Once the models have been
run, the function popan.derived is called with the list of model results which computes the abundances and related statistics (immigration BiGross) by model averaging
over the set of models.  These are then stored in the list cfdata which is returned with the correction factors by occasion (cfbyocc) and the marklist of final model results.
The cfdata is used by correct_dead and compute_cf functions to create estimates of total number of pups that died in years with no tagging data.

Note: In 1994 the number of untagged and stacked was not recorded on the last occasion; thus the correction factor for the last occasion is not useful.
}
\examples{
\dontrun{
smi2016.cu.popan.results=cu_popan.cf(2016)
save(smi2016.cu.popan.results,file="smi2016.cu.popan.results.rda")
}
}
\seealso{
correct_dead
}
\author{
Jeff Laake
}
