% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIPinnipedAnalysis-package.R
\docType{data}
\name{cufo}
\alias{cufo}
\title{Callorhinus frequency of occurrence data}
\format{A dataframe with fields 
\preformatted{
SCATNUM           - scat number id 
PREYSP            - prey species
FO_OT.BK          - landings of anchovy
FO_OTHER          - landings of mackerel
DATE              - landings of squid'
Year              - 
}}
\description{
A record for each species in a scat
}
\keyword{datasets}
