% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mortality.stats.R
\name{CuMortalityStats}
\alias{CuMortalityStats}
\alias{ZcMortalityStats}
\title{Early pup mortality estimation}
\usage{
CuMortalityStats(PreLiveCountCf=1.33,PostLiveCountCf=1.25,fdir1,fdir2,years=NULL)
       ZcMortalityStats(year,dead,pups)
}
\arguments{
\item{PreLiveCountCf}{Multiplicative correction factor for observed
mortality prior to live count to account for dead pups that were missed
and decomposed or were buried}

\item{PostLiveCountCf}{Multiplicative correction factor for observed
mortality after live count}

\item{fdir1}{database directory for CIPinnipedCensusQuery}

\item{fdir2}{database directory for CIPinnipedCensusMaster}

\item{years}{vector of years that should be selected if not all years}

\item{year}{year for mortality stats}

\item{dead}{is a dataframe of corrected dead pup counts across occasions}

\item{pups}{is the estimated number of pups produced in the mortality count area}
}
\value{
dataframe containing year,SurveyDate,Pups(number of pups alive at
  time),ObservedDead (number observed to die in that interval)
  AdjustedDead(corrected count that died),MortalityRate(mortality rate
  during the interval between surveys),DailyMortalityRate (mortality rate on
  a constant daily rate), CumS (cummulative survival to that time).
}
\description{
Creates a survivorship curve from dead pup count surveys.  The curve extends
from birth to either Aug or Sept-Oct depending on the extent of the
mortality surveys which has varied across the years.
}
\details{
The script EarlyPupMortality.r runs the functions to construct the
ZcEarlPupMortality and CuEarlyPupMortality tables in
CIPinnipedCensusQuery.mdb.  This function uses the ZcProduction and
CuProduction tables created by PupProduction scrupt and should be
run after it is updated with any new data. The CU values use pre-defined correction factors
whereas the ZC values use correction factor constructed from tagging data on dead pups. (see popan.cf)
}
\author{
Jeff Laake
}

