% Generated by roxygen2 (4.0.1): do not edit by hand
\name{do.pup.production}
\alias{do.pup.production}
\title{Pup production estimation}
\usage{
do.pup.production(fdir = NULL)
}
\arguments{
\item{fdir}{directory for data files; if NULL uses location specified in databases.txt of CalcurData package; if "" uses databases in CalcurData pacakge; otherwise uses specified directory location}
}
\value{
None
}
\description{
For Zc and Cu, creates data tables in the ACCESS CIPinnipedCensusQuery database with the early pup
mortality estimates during the season in each year.  Also constructs pdf
plots of those values.
}
\details{
Uses the \code{\link{production.stats}} to create ZcProduction and
CuProduction tables for SMI in CIPinnipedCensusQuery.mdb. Any expected
warnings or error messages are printed in PupProduction.log and unexpected
errors/coding problems would be found in PupProduction.out.  If there are
any errors then the production tables will not be created. In addition to
creating the tables, it also produces a set of plots of the results in pdfs.
}
\author{
Jeff Laake
}

