% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIPinnipedAnalysis-package.R
\docType{data}
\name{fo}
\alias{fo}
\title{Prey species ID data}
\format{A dataframe with fields 
\preformatted{
SCATNUM - scat identifier
PREYSP  - code for prey species in scat
FO_OT.BK - 1 if identified by beak or otolith
FO_OTHER - 1 if indentified by other structure
Date     - date scat collected
Year     - year of collection
}
Created with FO_DAT.txt file and following R code
fo=read.delim("Fo_DAT.txt",sep="\\t",header=TRUE,colClasses=c("factor","factor","character","character","NULL","character","NULL","NULL","numeric","NULL"))
#remove 1994 and 2007 because there is no prey species fo data
fo=droplevels(fo[!fo$Year%in%c(1994,2007),])
save(fo,file="fo.rda")}
\description{
Prey species identified from Zalophus scat data
}
\keyword{datasets}

