% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.cu.weights.r
\name{get.cu.weights}
\alias{get.cu.weights}
\title{Extracts weights for Cu pups
: from various tables in the cutagnew.mdb ACCESS
database and returns a dataframe with ancillary fields added.}
\usage{
get.cu.weights(fdir = NULL, ENYears = c(1976, 1983, 1984, 1986, 1987, 1992,
  1997, 1998, 2002, 2009))
}
\arguments{
\item{fdir}{directory for cutagnew.mdb}

\item{ENYears}{values of years that will be flagged as ENSO years}
}
\value{
dataframe of weights for Cu
}
\description{
Extracts weights for Cu pups
: from various tables in the cutagnew.mdb ACCESS
database and returns a dataframe with ancillary fields added.
}
\examples{
\donttest{
source(file.path(system.file(package="CIPinnipedAnalysis"),"CuWeightAnalysis.r"))
}
}
\author{
Jeff Laake
}

