% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attendance.R
\name{get_attendance}
\alias{get_attendance}
\title{Create daily attendance encounter history for female sea lions}
\usage{
get_attendance(resights, year = 2006, firstday = 20, lastday = 25,
  areas = c("OCV", "EACS", "LNC", "LTC"))
}
\arguments{
\item{resights}{dataframe of zalpohus resights from Alive table in database}

\item{year}{4 digit year to extract}

\item{firstday}{numeric value of first day in May to use}

\item{lastday}{numeric value of last day in July to use}

\item{areas}{area codes to filter resightings}
}
\value{
list with a vector encounter histories (ch), vector of days when sighting took place (seenDays) and vector of all days (Days). Length of ch will be max of Days.
}
\description{
For a given year and day range (firstday,lastday), finds females that have been seen with a pup in July 
in a set of areas (e.g., LTC,LNC,EACS,OCV) and constructs a daily encounter history of seen (1) and not seen(0)
to model attendance patterns.
}
\examples{
resights=getCalcurData(db="Zc", tbl="Alive")
get_attendance(resights,year=2006,firstday=5)

}
\author{
Jeff Laake
}
