% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fo_data.r
\name{get_fo}
\alias{get_fo}
\title{Computes Frequency of Occurrence (FO) for Diet Data}
\usage{
get_fo(fo, species, exclude = FALSE, years = NULL, labels = NULL)
}
\arguments{
\item{fo}{dataframe of prey species identified in each scat and whether it was identified by otolith/beak or other}

\item{species}{a vector of prey species codes for which FO is computed. For more than one species code
it is the FO of any of the species.}

\item{exclude}{if TRUE, then uses any species other than those listed in species}

\item{years}{a vector of cut points for year groupings}

\item{labels}{to be used for year groupings}
}
\value{
A list with vector of FO values for each year in the data and number of scats for each year
}
\description{
For a given set of species codes it computes FO by extracting the observed scats and
combines into year groupings. For years before 1992 when only otoliths were
used, it "corrects" based on data collected after 1991.
}
\details{
Dataframe scats no longer used because some had entries that were not identified so fo was too large.  Now the counts
of scats are calculated from unique number of scats in prey (fo). This means for the 1980s any scat that could have 
had bone only is no longer in the count, so the bone_cf is not used or fo values greater than one could be obtained.
Thus up through 1991, fo is based on otloliths and beaks only.
}
\author{
Jeff Laake
}
