% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdead_ch.r
\name{getdead_ch}
\alias{getdead_ch}
\title{Dead pup capture histories}
\usage{
getdead_ch(island, year, development = "fullterm", merge = TRUE)
}
\arguments{
\item{island}{("SMI" or "SNI")}

\item{year}{four digit numeric year}

\item{development}{either "premie" or "fullterm"}

\item{merge}{if TRUE, merges disparate area codes into "PTS","SCV","WCV","NWC"}
}
\value{
\preformatted{dataframe containing 
ch(capture history), 
Carcass condition: F-fresh, D- decomposing, P- pancake, 
Position: A (Above) or B (Below)
Substrate: N (non-consolidated - sandy), C (consolidated - rocky)
AreaCode - beach area"}
}
\description{
Creates capture histories for tagged and untagged dead pups for POPAN analysis
}
\details{
Extracts data for a particular year, island and pup development stage (premie/fullterm) from 
Zc dead tag initial and Zc dead tag resight for tagged dead pups, 
and Zc Cu dead pup census for untagged stacked dead pups. Creates a capture history
for each pup and uses -1 for freq for any stacked pup.  Includes initial substrate, carcass
condition and beach position as covariates. Note FP (flood pond) has been used infrequently
and has been coverted to A (above beach crest) versus B (below beach crest).
  
It reports any mismatches (id in resights with no initial) and any duplicate initial data
records.
}
\author{
Jeff Laake
}
