% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIPinnipedAnalysis-package.R
\docType{data}
\name{hake}
\alias{hake}
\title{Hake Biomass Estimates}
\format{A dataframe with fields 
\preformatted{
Year           - year 
Age0           - biomass of Age 0 hake
Age1           - biomass of Age 1 hake
Age2           - biomass of Age 2 hake
}}
\description{
A record for 1977-2015 from
Grandin, C.J., A.C. Hicks, A.M. Berger, A.M. Edwards, N. Taylor, I.G. Taylor, and S. Cox. 2016. Status of the Pacific Hake (whiting) stock in U.S. and Canadian waters in 2016. Prepared by the Joint Technical Committee of the U.S. and Canada Pacific Hake/Whiting Agreement, National Marine Fisheries Service and Fisheries and Oceans Canada. 165 p.
}
\keyword{datasets}
