% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIPinnipedAnalysis-package.R
\docType{data}
\name{Landings}
\alias{Landings}
\title{CALCOFI Fisheries Landings}
\format{A dataframe with fields 
\preformatted{
Year           - year 
Sardine        - landings of sardine
Anchovy        - landings of anchovy
Mackerel       - landings of mackerel
Squid          - landings of squid
}}
\description{
A record for 1975-2014
}
\keyword{datasets}
