% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIPinnipedAnalysis-package.R
\docType{data}
\name{loligo}
\alias{loligo}
\title{Loligo Quarterly Biomass Estimates}
\format{A dataframe with fields 
\preformatted{
Year           - year 
Quarter        - quarter of year
SSBmt          - stock biomass estimate in metric tons
}}
\description{
From region 2, using M=0.15 Table 2 from
Dorval, E., P. R. Crone, and J. D. McDaniel. 2013. Variability of egg escapement, fishing mortality and spawning population in the market squid fishery in the California Current Ecosystem. Mar. Freshw. Res. 64:80-90.
A record for 1999-2006
}
\keyword{datasets}

