% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality.stats.R
\name{MortalityStats}
\alias{MortalityStats}
\title{Early pup mortality estimation}
\usage{
MortalityStats(year, dead, pups)
}
\arguments{
\item{year}{year for mortality stats}

\item{dead}{is a dataframe of corrected dead pup counts across occasions}

\item{pups}{is the estimated number of pups produced in the mortality count area}
}
\value{
dataframe containing year,Pups(number of pups alive at
  time),Numbersead (number estimated to die in that interval)
  DaysFrom15June,MortalityRate(mortality rate
  during the interval between surveys),DailyMortalityRate (mortality rate on
  a constant daily rate), CumS (cummulative survival to that time).
}
\description{
Creates a survivorship curve from dead pup count surveys.  The curve extends
from birth to either Aug or Sept-Oct depending on the extent of the
mortality surveys which has varied across the years.
}
\details{
The script EarlyPupMortality.r runs the functions to construct the
ZcEarlPupMortality and CuEarlyPupMortality tables in
CIPinnipedCensusQuery.mdb.  This function uses the ZcProduction and
CuProduction tables created by PupProduction scrupt and should be
run after it is updated with any new data. The ZC and CU values use correction factors 
constructed from tagging data on dead pups. (see popan.cf and cu_popan.cf)
}
\author{
Jeff Laake
}
