% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popan.cf.r
\name{popan.cf}
\alias{popan.cf}
\title{Construct correction factor data from dead pup tagging studies}
\usage{
popan.cf(island, year, silent = TRUE, area = FALSE)
}
\arguments{
\item{island}{("SMI" or "SNI")}

\item{year}{four digit numeric year}

\item{silent}{if TRUE shows each model as it is run and any problems that occur; if FALSE this is hidden each survey date (occasion); if non-null provides correction at that date}

\item{area}{if TRUE, uses area in place of position and substrate to fit models for tagging data >1997}
}
\value{
a list containing a marlist of the final model results, a list with the correction factor data (cfdata) and correction factors by occasion (cfbyocc).
}
\description{
For a given island (SMI or SNI) and year, the dead pup tagging data is analyzed with a set of POPAN 
models to create correction factors for the observed dead pup counts to estimate the total number that died.
}
\details{
The function getdead_ch is used to extract the data from the CIPinnipedCensusMaster database and 
create capture histories for tagged and untagged (stacked) pups. This is then analyzed with the POPAN model in MARK using
the RMark package.  A set of models is examined that depends on the year because the number of sampling occasions
and the data collected has varied.  Prior to 1998, the position on the beach and substrate were not recorded for
dead pups when they were stacked.  However, the area was recorded and the substrate is largely similar in each area. Thus
for years prior to 1998 the area was used as a grouping factor and in 1998 and beyond, substrate and position were used as
grouping factors. In POPAN models, p on the first and last occasion are not separately estimable in a time model
so these have been restricted such that p for occasion 1 is the same as occasion 2 and p for the last occasion is the same as p for the 
next to last occasion (see definition of time2).

Each set of models is run and then they are re-run using the initial values from the best model.  This helps with
model convergence which can fail for POPAN models.  To see which models are failing set silent=FALSE. Once the models have been
run, the function popan.derived is called with the list of model results which computes the abundances and related statistics (immigration BiGross) by model averaging
over the set of models.  These are then stored in the list cfdata which is returned with the correction factors by occasion (cfbyocc) and the marklist of final model results.
The cfdata is used by correct_dead and get_cf functions to create estimates of total number of pups that died in years with no tagging data.

Note: In 1994 the number of untagged and stacked was not recorded on the last occasion; thus the correction factor for the last occasion is not useful.
}
\examples{
\dontrun{
#note this will construct all of the correction factor data; it will take awhile
smi1994.popan.results=popan.cf("SMI",1994)
smi1995.popan.results=popan.cf("SMI",1995)
smi1998.popan.results=popan.cf("SMI",1998)
smi2002.popan.results=popan.cf("SMI",2002)
sni2006.popan.results=popan.cf("SNI",2006)
smi1998a.popan.results=popan.cf("SMI",1998,area=TRUE)
smi2002a.popan.results=popan.cf("SMI",2002,area=TRUE)
}
}
\author{
Jeff Laake
}
\seealso{
correct_dead
}

