% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIPinnipedAnalysis-package.R
\docType{data}
\name{Rockfish}
\alias{Rockfish}
\title{Rockfish Biomass Estimates}
\format{A dataframe with fields 
\preformatted{
Year           - year 
entomelas      - recruitment estimate/1000 of the species
goodei         - recruitment estimate/1000 of the species
jordani        - recruitment estimate/1000 of the species
paucispinis    - recruitment estimate/1000 of the species
pinniger       - recruitment estimate/1000 of the species
}}
\description{
A record for 1980-2010 from Table 7 of
Ralston, S., K. M. Sakuma, and J. C. Field. 2013. Interannual variation in pelagic juvenile rockfish (Sebastes spp.) abundance - going with the flow. Fish. Oceanogr. 22:288-308.
}
\keyword{datasets}
