% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIPinnipedAnalysis-package.R
\docType{data}
\name{sardine}
\alias{sardine}
\title{Sardine Biomass Estimates}
\format{A dataframe with fields 
\preformatted{
Year           - year 
Age            - either age 0 or age 1
Abundance      - estimate of abundance at age
Biomass        - estimate of stock biomass at age for the year
}}
\description{
A record for 1981-2013 of age 0 and age 1 sardines from
Hill, K. T., P. R. Crone, E. Dorval, and Macewicz. 2015. ASSESSMENT OF THE PACIFIC SARDINE RESOURCE IN 2015 FOR U.S.A. MANAGEMENT IN 2015-16. NOAA-TM-NMFS-SWFSC-546 .
}
\keyword{datasets}
