% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CIPinnipedAnalysis-package.R
\docType{data}
\name{scats}
\alias{scats}
\title{Scat record data}
\format{A dataframe with fields
\preformatted{
SCATNUM - scat identifier
SPCOD   - code for pinniped species
DATE    - date scat collected
OTO     - yes of TRUE if any otloliths in scat
BKS     - yes or TRUE if any beaks in scat
BNE     - yes or TRUE if any bone in scat
Year     - year of collection
Month    - month of collection
}
Created with scats.txt file and the collowing R code
scats=read.delim("scats.txt",sep="\\t",header=TRUE,colClasses=c("factor","factor","character","NULL","NULL","NULL","NULL","character","character","character","numeric","numeric"))
#remove 1994 and 2007 because there is no prey species fo data
scats=scats[!scats$Year%in%c(1994,2007),]
# exclude any with all FALSE or NO for OTO,BKS,BNE except those in include which have prey data
xx=merge(scats,fo)
include=xx[(toupper(xx$BNE)%in%c("FALSE","NO")&toupper(xx$OTO)%in%c("FALSE","NO")&toupper(xx$BKS)%in%c("FALSE","NO"))&!is.na(xx$PREYSP),"SCATNUM"]
rm(xx)
scats=scats[toupper(scats$BNE)%in%c("YES","TRUE")|toupper(scats$BKS)%in%c("YES","TRUE")|toupper(scats$OTO)%in%c("YES","TRUE")|scats$SCATNUM%in%include,]
scats=droplevels(scats)
save(scats,file="scats.rda")}
\description{
A record for each scat sampled
}
\keyword{datasets}

