% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIPinnipedAnalysis-package.R
\docType{data}
\name{SeaLevelHeight}
\alias{SeaLevelHeight}
\title{Port San Luis Sea Level Height Data}
\format{A dataframe with fields 
\preformatted{
Year - year of data measurement
Month - numeric month of measurement
SeaLevelHeight - monthly mean sea level height in meters with seasonal fluctuations removed
}}
\description{
A sequence of monthly average sea level height at Port San Luis
http://tidesandcurrents.noaa.gov/sltrends/sltrends_station.shtml?stnid=9412110
}
\keyword{datasets}

