% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/store_weights.r
\name{store_weights}
\alias{store_weights}
\title{Store weights in CIPinnipedCensusQuery}
\usage{
store_weights(x, species = "Zc", fdir = NULL)
}
\arguments{
\item{x}{dataframe containing weights to be stored in table}

\item{species}{species code; either Zc or Cu}

\item{fdir}{directory for databases; if NULL uses location from databases.txt in CalcurData package; if "" uses databases stored in package directory; if neither you can specify the location}
}
\value{
None
}
\description{
Stores standardized weights into table in ACCESS database for Zc and Cu
depending on the value of species.
}
\author{
Jeff Laake
}

