% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIPinnipedAnalysis-package.R
\docType{data}
\name{strandingAL}
\alias{strandingAL}
\title{Zalophus stranding data from Marine Mammal Center}
\format{A dataframe with fields 
\preformatted{
Name           - sea lion identifier 
Age            - sea lion age in years
Sex            - sex of sea lion M (male) or F (Female) 
Length         - standard length in cm
}}
\description{
Provided by Denise Greig from her stranding paper
Greig, D.J., F.M.D. Gulland, and C. Kreuder. 2005. A decade of live California sea lion
(Zalophus californianus) strandings along the central California coast: Causes and trends,
1991-2000. Aquat. Mammals 31: 11-22.
}
\keyword{datasets}
