% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIPinnipedAnalysis-package.R
\docType{data}
\name{uwi33}
\alias{uwi33}
\title{Daily Upwelling Index values at 33N
http://www.pfeg.noaa.gov/products/PFEL/modeled/indices/upwelling/NA/data_download.html}
\format{A dataframe with fields 
\preformatted{
YYYYMMDD    - 4 digit year, 2 digit month and 2 digit day
Index       - upwelling index for the day 
}}
\description{
A record for from 1967 onwards
}
\keyword{datasets}
