% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permit_report.r
\name{ZCpermit_report}
\alias{ZCpermit_report}
\title{Create Zc Permit report}
\usage{
ZCpermit_report(year, dir = NULL)
}
\arguments{
\item{year}{numeric year for permit work}

\item{dir}{directory for location of files; dir=NULL uses default value in databases.txt in CalcurData}
}
\description{
Produces a table with tally of sampled animals for permit
}
\examples{
ZCpermit_report(2013)

}
\author{
Jeff Laake
}

