% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{CIPinnipedAnalysis-package}
\alias{CIPinnipedAnalysis}
\alias{CIPinnipedAnalysis-package}
\title{CIPinnipedAnalysis}
\description{
Channel Islands Pinniped Analysis functions
: pup production, early pup mortality, SST and other enviromental measurement computations
}
\details{
\tabular{ll}{ Package: \tab CIPinnipedAnalysis\cr Type: \tab Package\cr
Version: \tab 1.0\cr Date: \tab 2009-11-30\cr License: \tab GPL-2\cr
LazyLoad: \tab yes\cr }
}
\examples{
# Note this uses default argument value of NULL for dir which
# means that the locations for the databases is specified by databases.txt in CalcurData package
check.dead.pup.areas()
source(file.path(system.file(package="CIPinnipedAnalysis"),"PupProduction.r"))
source(file.path(system.file(package="CIPinnipedAnalysis"),"EarlyPupMortality.r"))
}
\author{
Jeff Laake

Maintainer: <jeff.laake@noaa.gov>
}
\keyword{package}

