% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.live.dead.R
\name{construct.live}
\alias{construct.dead}
\alias{construct.live}
\title{Constructs dataframes of live or dead pup counts}
\usage{
construct.live(island, species, dir = NULL)
}
\arguments{
\item{island}{Either "SMI" or "SNI"}

\item{species}{Either "Zc" or "Cu"}

\item{dir}{directory for data files; if NULL uses location specified in databases.txt of CalcurData package; if "" uses databases in CalcurData pacakge; otherwise uses specified directory location}
}
\value{
for \code{construct.live} a dataframe containing YearArea, Survey
  date, Area, AvgCount, Year, Month, Day, DeadPupArea (Y/N if included in
  area sampled for dead pups).

for \code{construct.dead} a dataframe containing YearArea, SurveyNumber,
  SurveyDate, Area, count, Year, Month, Day, DeadPupArea (Y/N if included in
  area sampled for dead pups), and MatchingLiveArea.
}
\description{
Extracts data from CIPinnipedMaster database and constructs dataframe of
dead pup counts or live pup counts for specified island and species.
}
\details{
\code{construct.live} extracts data from table \code{Zc Cu live pup census}
and sums counts for an observer within an area for a year and then averages
those counts across observers that counted the same area that year.

\code{construct.dead} extracts data from table \code{Zc Cu dead pup census}
and table \code{Zc dead tag initial} and totals dead pups by survey
number/date.  It will produce errors if there are more than one survey date
for each survey number within an area in a year.
}
\author{
Jeff Laake
}

