% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_dead.r
\name{correct_dead}
\alias{average_cf}
\alias{compute_cf}
\alias{correct_dead}
\alias{process_ch}
\title{Correction factors for dead pup counts}
\usage{
correct_dead(island,year)
       average_cf(island,year,ndays1=Inf,ndays2,area=NULL,lev=NULL)
       compute_cf(BiGross,ndays1=Inf,ndays2)
       process_ch(ch,freq=NULL,all=FALSE)
}
\arguments{
\item{island}{("SMI" or "SNI")}

\item{year}{four digit numeric year}

\item{BiGross}{is the estimates of gross immigration into the population of dead pups; it is a list with results from fitted models}

\item{ndays1}{number of days from 1 July for beginning of interval; -Inf default which starts at beginning}

\item{ndays2}{number of days from 1 July at end of interval (time of pup count)}

\item{area}{for computing correction factor with tagging data prior to 1998}

\item{lev}{position on beach and substrate (AC,AN,BC,BN) for computing correction factor with data in 1998 and later}

\item{ch}{vector of capture histories}

\item{freq}{frequency of capture history}

\item{all}{logical indicating whether all quantities should be computed from capture history}
}
\value{
correct_dead returns a list containing a dataframe with corrected dead pup counts by occasion and strata (eg group - area or substrate/position) and a dataframe with a total across strata by occasion. 
average_cf returns an average correction factor for a particular range of dates by interpolation and then averaging over years (cfyears) for a particular area code or postion-substrate strata.
}
\description{
Correction factors for a set of dead pup counts are created for each survey date or a single date from 
estimates created from POPAN model (cfdata) on tagging data.
}
\details{
For a given island and year, it calls getdead_ch  and process_ch to create the capture history for the dead pup data and then 
uses process.data from RMark to process the capture history data for the POPAN model. If cfyear > 1997 it uses Position and Substrate as group factors and if 
cfyear < 1997 it uses Area code which is split into 4 areas: PTS(NEP,NWP,PBP),SCV (SCV,PBS,EAC and WAC), WCV and NWC. The argument cfdata
contains the estimates from the POPAN model to create the correction factor. It is created by the function popan.cf which runs a set of POPAN models
with the tagging data from year yyyy (1994,1995,1998,2002 for SMI and 2006 for SNI) and creates smiyyyy.popan.results or sniyyyypopan.results with the model results and model averaged values for cfdata. If ndays is null it creates a correction factor for each 
survey occasion. If ndays, number of days from 1 July, is not null then it computes with interpolation the number observed dead at that date and
its correction factor.

The function average_cf computes an average correction factor for a set of years with tagging data. This is only useful at present with SMI becuse SNI has only 
one year of tagging data but it can be used with SNI as well.  It can use an area (SCV,PTS,WCV,NWC) or lev value (AC,AN,BC,BN) to compute specific correction 
factors. Based on the island and year, it knows which set of years to use for the correction factor.

Note: In 1994 the number of untagged and stacked was not recorded on the last occasion; thus the correction factor for the last occasion is not useful.
}
\author{
Jeff Laake
}
\seealso{
popan.cf getdead_ch
}

