% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get.zc.weights.r
\name{get.zc.weights}
\alias{get.zc.weights}
\title{Extracts weights for Zc pups
:from various tables in the BrandMaster.mdb ACCESS database and returns a dataframe with ancillary fields added.}
\usage{
get.zc.weights(fdir = NULL, ENYears = c(1976, 1983, 1984, 1986, 1987, 1992,
  1997, 1998, 2002, 2009), exclude = TRUE)
}
\arguments{
\item{fdir}{directory for data files; if NULL uses location specified in databases.txt of CalcurData package; if "" uses databases in CalcurData pacakge; otherwise uses specified directory location}

\item{ENYears}{values of years that will be flagged as ENSO years}

\item{exclude}{if TRUE, it excludes any records with missing weight value; setting exclude=FALSE extracts all records which is useful for creating age-length key}
}
\value{
dataframe of weights for Zc
}
\description{
Extracts weights for Zc pups
:from various tables in the BrandMaster.mdb ACCESS database and returns a dataframe with ancillary fields added.
}
\examples{
# following runs adjustment model and all subsequent growth models
# and then stores table in CIPinnipedCensusQuery database. This will take
# a substantial amount of time
source(file.path(system.file(package="CIPinnipedAnalysis"),"ZcWeightAnalysis.r"))
}
\author{
Jeff Laake
}

