% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.zc.weights}
\alias{get.zc.weights}
\title{Extracts weights for Zc pups
:from various tables in the BrandMaster.mdb ACCESS database and returns a dataframe with ancillary fields added.}
\usage{
get.zc.weights(fdir = NULL, ENYears = c(1976, 1983, 1984, 1986, 1987, 1992,
  1997, 1998, 2002, 2009))
}
\arguments{
\item{fdir}{directory for data files; if NULL uses location specified in databases.txt of CalcurData package; if "" uses databases in CalcurData pacakge; otherwise uses specified directory location}

\item{ENYears}{values of years that will be flagged as ENSO years}
}
\value{
dataframe of weights for Zc
}
\description{
Extracts weights for Zc pups
:from various tables in the BrandMaster.mdb ACCESS database and returns a dataframe with ancillary fields added.
}
\examples{
source(file.path(system.file(package="CIPinnipedAnalysis"),"ZcWeightAnalysis.r"))
}
\author{
Jeff Laake
}

